Adds option to hide regattas from front page
Adds a "Hide From Front Page" option to regattas, allowing administrators to exclude them from the main display.

This can be useful for hiding regattas that are not yet ready for public viewing, or that are no longer relevant.

Adds a migration to add the new property.
Updates the regatta service to filter out hidden regattas.